requirements.txt
aiosqlite
pyyaml
pycryptodome
tornado=6.0.4
